requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
emoji==2.8.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
proglog==0.1.10
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
starlette==0.31.0
tqdm==4.66.1
urllib3==2.0.4
uvicorn==0.23.2
